feat: add and expose ExecutableCommand
